moved fixtures to validation_test package to prevent accidental use in normal package
